# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PostMem API",
    "description": "个人知识库系统 API - 支持本地嵌入向量和智能文本切割（自动生成）",
    "version": "1.0.0",
    "contact": {
      "name": "PostMem"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "当前服务器"
    }
  ],
  "paths": {
...
    "/api/models/{id}": {
      "get": {
        "summary": "单个模型 API",
        "description": "单个模型 API",
        "tags": [
          "Models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "模型ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Model"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Model"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "单个模型 API",
        "description": "单个模型 API",
        "tags": [
          "Models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "模型ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Model"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Model"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModelRequest"
              },
              "example": {}
            }
          }
        }
      },
      "delete": {
        "summary": "单个模型 API",
        "description": "单个模型 API",
        "tags": [
          "Models"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "模型ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Model"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Model"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/providers/{id}": {
      "get": {
        "summary": "单个提供商 API",
        "description": "单个提供商 API",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "提供商ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Provider"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Provider"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "单个提供商 API",
        "description": "单个提供商 API",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "提供商ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Provider"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Provider"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderRequest"
              },
              "example": {}
            }
          }
        }
      },
      "delete": {
        "summary": "单个提供商 API",
        "description": "单个提供商 API",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "提供商ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Provider"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Provider"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vendors/{id}": {
      "get": {
        "summary": "单个厂商 API",
        "description": "单个厂商 API",
        "tags": [
          "Vendors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "厂商ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Vendor"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Vendor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "单个厂商 API",
        "description": "单个厂商 API",
        "tags": [
          "Vendors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "厂商ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Vendor"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Vendor"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVendorRequest"
              },
              "example": {}
            }
          }
        }
      },
      "delete": {
        "summary": "单个厂商 API",
        "description": "单个厂商 API",
        "tags": [
          "Vendors"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "厂商ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Vendor"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Vendor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Vendor": {
        "type": "object",
        "required": [
          "id",
          "name",
          "url",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "id"
          },
          "name": {
            "type": "string",
            "description": "name"
          },
          "url": {
            "type": "string",
            "description": "url"
          },
          "chatModelClass": {
            "type": "string",
            "enum": [
              "string",
              "null"
            ],
            "description": "chatModelClass"
          },
          "embeddingModelClass": {
            "type": "string",
            "enum": [
              "string",
              "null"
            ],
            "description": "embeddingModelClass"
          },
          "factoryCode": {
            "type": "string",
            "enum": [
              "string",
              "null"
            ],
            "description": "factoryCode"
          },
          "isActive": {
            "type": "boolean",
            "description": "isActive"
          },
          "createdAt": {
            "type": "string",
            "description": "createdAt"
          },
          "updatedAt": {
            "type": "string",
            "description": "updatedAt"
          }
        }
      },
      "Provider": {
        "type": "object",
        "required": [
          "id",
          "name",
          "vendorId",
          "baseUrl",
          "config",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "id"
          },
          "name": {
            "type": "string",
            "description": "name"
          },
          "vendorId": {
            "type": "string",
            "description": "vendorId"
          },
          "vendor": {
            "$ref": "#/components/schemas/Vendor",
            "description": "vendor"
          },
          "apiKey": {
            "type": "string",
            "description": "apiKey"
          },
          "baseUrl": {
            "type": "string",
            "description": "baseUrl"
          },
          "config": {
            "type": "object",
            "description": "config"
          },
          "isActive": {
            "type": "boolean",
            "description": "isActive"
          },
          "createdAt": {
            "type": "string",
            "description": "createdAt"
          },
          "updatedAt": {
            "type": "string",
            "description": "updatedAt"
          }
        }
      },
      "Model": {
        "type": "object",
        "required": [
          "id",
          "providerId",
          "name",
          "modelType",
          "config",
          "isActive",
          "isDefault",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "id"
          },
          "providerId": {
            "type": "string",
            "description": "providerId"
          },
          "name": {
            "type": "string",
            "description": "name"
          },
          "displayName": {
            "type": "string",
            "description": "displayName"
          },
          "modelType": {
            "type": "string",
            "description": "modelType"
          },
          "config": {
            "type": "object",
            "description": "config"
          },
          "isActive": {
            "type": "boolean",
            "description": "isActive"
          },
          "isDefault": {
            "type": "boolean",
            "description": "isDefault"
          },
          "createdAt": {
            "type": "string",
            "description": "createdAt"
          },
          "updatedAt": {
            "type": "string",
            "description": "updatedAt"
          }
        }
      },
...
      "UpdateProviderRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name"
          },
          "vendorId": {
            "type": "string",
            "description": "vendorId"
          },
          "apiKey": {
            "type": "string",
            "description": "apiKey"
          },
          "baseUrl": {
            "type": "string",
            "description": "baseUrl"
          },
          "config": {
            "type": "object",
            "description": "config"
          },
          "isActive": {
            "type": "boolean",
            "description": "isActive"
          }
        }
      },
...
      "UpdateVendorRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name"
          },
          "chatModelClass": {
            "type": "string",
            "description": "chatModelClass"
          },
          "embeddingModelClass": {
            "type": "string",
            "description": "embeddingModelClass"
          },
          "factoryCode": {
            "type": "string",
            "description": "factoryCode"
          },
          "isActive": {
            "type": "boolean",
            "description": "isActive"
          }
        }
      },
...
      "UpdateModelRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name"
          },
          "displayName": {
            "type": "string",
            "description": "displayName"
          },
          "modelType": {
            "type": "string",
            "description": "modelType"
          },
          "config": {
            "type": "object",
            "description": "config"
          },
          "isActive": {
            "type": "boolean",
            "description": "isActive"
          },
          "isDefault": {
            "type": "boolean",
            "description": "isDefault"
          }
        }
      },
...
    }
  },
...
}